Try Catch Erreurs

+++ src/Controller/APIEquipesController.php
@@ -18,41 +18,83 @@ use App\Repository\EquipesRepository;
 
 use App\Entity\Equipes;
 
+// Réponses Json d'Erreur
+define("ERR_ID_INTROUVABLE", new JsonResponse(
+    json(["status" => "Bad Request", "details" => "Aucun id spécifié"]),
+    Response::HTTP_BAD_REQUEST,
+    [], true
+));
+
+define("ERR_ENTITE_INTROUVABLE", new JsonResponse(
+    json(["status" => "Not Found", "details" => "Entité introuvable"]),
+    Response::HTTP_NOT_FOUND,
+    [], true
+));
+
+
 #[Route('/api/equipes')]
 class APIEquipesController extends AbstractController
 {
     #[Route('', name: 'api_get', methods: ['GET'])]
     public function getEquipeId(Request $request, EquipesRepository $equipesRepository, SerializerInterface $serializer): Response
     {
-        (int) $id = $request->query->get('id');
-
         if (isset($id)) {
+            (int) $id = $request->query->get('id');
             $equipe = $equipesRepository->find($id);
-            $equipeJson = $serializer->serialize($equipe, 'json', ['groups' => 'equipe']);
 
-            return new JsonResponse($equipeJson, Response::HTTP_OK, [], true);
+            if ($equipe === null) { return ERR_ENTITE_INTROUVABLE; }
+
+            $equipesJson = $serializer->serialize($equipe, 'json', ['groups' => 'equipe']);
+
         } else {
             $equipes = $equipesRepository->findAll();
             $equipesJson = $serializer->serialize($equipes, 'json', ['groups' => 'equipesMinimum']);
-
-            return new JsonResponse($equipesJson, Response::HTTP_OK, [], true);
         }
+
+        return new JsonResponse($equipesJson, Response::HTTP_OK, [], true);
     }
 
     #[Route('', name: 'delete_equipe', methods: ['DELETE'])]
-    public function delete(Request $request, EntityManagerInterface $entityManager, EquipesRepository $equipesRepository): Response
+    public function delete(Request $request, EntityManagerInterface $entityManager, EquipesRepository $equipesRepository): JsonResponse
     {
-        (int) $id = $request->query->get('id');
-        $equipe = $equipesRepository->find($id);
+        if (!isset($id)) { return ERR_ID_INTROUVABLE; }
+
+        try {
+            (int) $id = $request->query->get('id');
+            $equipe = $equipesRepository->find($id);
+        } catch (Exception $e) {
+            return new JsonResponse(
+                json([
+                    "status" => "Internal Server Error",
+                    "details" => $e
+                ]),
+                Response::HTTP_INTERNAL_SERVER_ERROR,
+                [], true
+            );
+        }
+
+        if ($equipe === null) { return ERR_ENTITE_INTROUVABLE; }
 
         $entityManager->remove($equipe);
         $entityManager->flush();
 
         if (null !== $equipesRepository->find($id)) {
-            return new Response("", Response::HTTP_INTERNAL_SERVER_ERROR, [], true);
+            return new JsonResponse(
+                json([
+                    "status" => "Internal Server Error",
+                    "details" => "L'entité n'a pas pu être supprimée"
+                ]),
+                Response::HTTP_INTERNAL_SERVER_ERROR,
+                [], true);
         }
 
-        return new Response("", Response::HTTP_NO_CONTENT, [], true);
+        return new JsonResponse(
+            json([
+                "status" => "No Content",
+                "details" => "L'entité a été supprimée"
+            ]),
+            Response::HTTP_NO_CONTENT,
+            [], true);
     }
 
 
@@ -62,16 +104,34 @@ class APIEquipesController extends AbstractController
         EntityManagerInterface $entityManager,
         EquipesRepository $equipesRepository,
         SerializerInterface $serializer
-        ): Response
+        ): JsonResponse
     {
-        $json = $request->getContent();
-        $equipe = $serializer->deserialize($json, 'App\Entity\Equipes', "json");
+        try {
+            $json = $request->getContent();
+            $equipe = $serializer->deserialize($json, 'App\Entity\Equipes', "json");
+        } catch (Exception $e) {
+            return new JsonResponse(
+                json([
+                    "status" => "Internal Server Error",
+                    "details" => $e
+                ]),
+                Response::HTTP_INTERNAL_SERVER_ERROR,
+                [], true
+            );
+        }
 
         $entityManager->persist($equipe);
         $entityManager->flush();
 
-        // return new Response(Response::HTTP_CREATED, Response::HTTP_CREATED, [], true);
-        return new JsonResponse($equipe->getNom(), Response::HTTP_OK, [], true);
+
+        return new JsonResponse(
+            json([
+                "status" => "OK",
+                "details" => "Entitée créée avec succès"
+            ]),
+            Response::HTTP_OK,
+            [], true
+        );
     }
 
 
@@ -85,34 +145,46 @@ class APIEquipesController extends AbstractController
     {
         $json = json_decode($request->getContent(), true);
 
-        
-        $equipe = $repo->find($json['id']);
+        if (!isset($json['id'])) { return ERR_ID_INTROUVABLE; }
 
-        
-        if ($equipe === null) {
+        try {
+            $equipe = $repo->find($json['id']);
+        } catch (Exception $e) {
             return new JsonResponse(
                 json([
-                    "error" => "Entité introuvable"
+                    "status" => "Internal Server Error",
+                    "details" => $e
                 ]),
-                Response::HTTP_NOT_FOUND,
-                [],
-                true
+                Response::HTTP_INTERNAL_SERVER_ERROR,
+                [], true
             );
         }
 
-        $serializer->deserialize(
-            $request->getContent(),
-            'App\Entity\Equipes',
-            "json",
-            [
-                AbstractNormalizer::OBJECT_TO_POPULATE => $equipe
-            ]
-        );
+        if ($equipe === null) { return ERR_ENTITE_INTROUVABLE; }
+
+        try {
+            $serializer->deserialize(
+                $request->getContent(),
+                'App\Entity\Equipes',
+                "json",
+                [AbstractNormalizer::OBJECT_TO_POPULATE => $equipe]
+            );
+        } catch (Exception $e) {
+            return new JsonResponse(
+                json([
+                    "status" => "Internal Server Error",
+                    "details" => $e
+                ]),
+                Response::HTTP_INTERNAL_SERVER_ERROR,
+                [], true
+            );
+        }
 
         $em->persist($equipe);
         $em->flush();
 
         $equipeJson = $serializer->serialize($equipe, 'json', ['groups' => 'equipe']);
+
 
         return new JsonResponse($equipeJson, Response::HTTP_OK, [], true);
     }
